resolve remaining warnings

+++ src/leveldb/db/version_edit.cc
@@ -116,7 +116,7 @@ static bool GetInternalKey(Slice* input, InternalKey* dst) {
 static bool GetLevel(Slice* input, int* level) {
   uint32_t v;
   if (GetVarint32(input, &v) &&
-      v < config::kNumLevels) {
+      v < static_cast<uint32_t>(config::kNumLevels)) {
     *level = v;
     return true;
   } else {
